% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.resize.R
\name{boot.resize}
\alias{boot.resize}
\title{Bootstrapping with Resized Sample}
\usage{
boot.resize(id, group1, group2, data, level, increase = 50)
}
\arguments{
\item{id}{character of level 1 grouping variable/identifier}

\item{group1}{character of level 2 grouping variable (ex: classrooms)}

\item{group2}{character of level 3 grouping variable (ex: schools), default is FALSE in case of only 2 levels of nesting.}

\item{data}{dataset from which to bootstrap}

\item{level}{character of level variable at which you want to increase sample size}

\item{increase}{= numerical value of multiplicative increase in sample size, default is 2}
}
\description{
Increase your sample size at level 1 or level 2 by bootstrapping. This function is intended for use with pilot study data to understand how great of an increase in sample size, whether through level 1 or level 2, would result in acceptable power for a full-scale study.
}
\examples{
boot.resize(id = "ID", group1 = "schoolid", group2 = FALSE, data = dat)
}
