% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmlm.R
\name{bootmlm}
\alias{bootmlm}
\title{Multi-level Bootstrapping}
\usage{
bootmlm(id, group1, group2 = FALSE, data)
}
\arguments{
\item{id}{character of level 1 grouping variable/identifier}

\item{group1}{character of level 2 grouping variable (ex: classrooms)}

\item{group2}{character of level 3 grouping variable (ex: schools), default is FALSE in case of only 2 levels of nesting.}

\item{data}{dataset from which to bootstrap}
}
\description{
Bootstrapping for 2- or 3-level multi-level data
}
\examples{
bootmlm(id = "ID", group1 = "classid", group2 = "schoolid", data = dat)
}
