% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.curve.R
\name{pow.curve}
\alias{pow.curve}
\title{Power Curve with Varying Alphas}
\usage{
boot.power(model, n, id, group1, group2, data)
}
\arguments{
\item{model}{vector containing the lmer formula}

\item{n}{number of bootstrap samples from which to run the model}

\item{id}{character of level 1 grouping variable/identifier}

\item{group1}{character of level 2 grouping variable (ex: classrooms)}

\item{group2}{character of level 3 grouping variable (ex: schools), default is FALSE in case of only 2 levels of nesting.}

\item{data}{dataset from which to bootstrap}
}
\value{
Returns power curves for all fixed parameters specified in the model.
}
\description{
Obtain a Power Curve based on bootstrapping with various alpha cut-offs. Displays the relation between power and Type I error rate.
}
\examples{
boot.power(model = Y ~ X + (1|schoolid), n = 1000, id = "ID", group1 = "schoolid", group2 = FALSE, data = dat)
}
