% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.resize.R
\name{power.resize}
\alias{power.resize}
\title{Power Estimate with Resized Sample}
\usage{
power.resize(model, n, id, group1, group2, data, level, increase = 50)
}
\arguments{
\item{model}{vector containing the lmer formula}

\item{n}{number of bootstrap samples from which to run the model}

\item{id}{character of level 1 grouping variable/identifier}

\item{group1}{character of level 2 grouping variable (ex: classrooms)}

\item{group2}{character of level 3 grouping variable (ex: schools), default is FALSE in case of only 2 levels of nesting.}

\item{data}{dataset from which to bootstrap}

\item{level}{character of level variable at which you want to increase sample size}

\item{increase}{= numerical value of multiplicative increase in sample size, default is 2}
}
\description{
Obtain a single power estimate based on bootstrapping from an increased sample size based on your data.
}
\examples{
power.resize(model = Y ~ X + (1|schoolid), n = 1000, id = "ID", group1 = "schoolid", group2 = FALSE, data = dat)
}
